+++ server/server.go
@@ -39,7 +39,7 @@ func StartServer(config *viper.Viper, flags map[string]interface{}) {
 	router.POST("/upload", handler.UploadApp)
 	router.POST("/update", handler.UpdateApp)
 	router.GET("/search", handler.GetAppByName)
-	router.DELETE("/deleteApp", handler.DeleteApp)
+	router.DELETE("/apps/delete", handler.DeleteSpecificVersionOfApp)
 	router.POST("/createChannel", handler.CreateChannel)
 	router.GET("/listChannels", handler.ListChannels)
 	router.DELETE("/deleteChannel", handler.DeleteChannel)
@@ -49,6 +49,9 @@ func StartServer(config *viper.Viper, flags map[string]interface{}) {
 	router.POST("/createArch", handler.CreateArch)
 	router.GET("/listArchs", handler.ListArchs)
 	router.DELETE("/deleteArch", handler.DeleteArch)
+	router.POST("/createApp", handler.CreateApp)
+	router.GET("/listApps", handler.ListApps)
+	router.DELETE("/deleteApp", handler.DeleteApp)
 
 	// get the port from the configuration file
 	port := config.GetString("PORT")