+++ CHANGELOG.md
@@ -2,14 +2,19 @@
 
 ## v1.3.4
 
+### Important Notes
+- **Important**: The route for deleting a specific app by ID has been renamed to  `/apps/delete` from `/deleteApp`.
+- **Important**: `/deleteApp` is now a new function.
+
 ### Features
 - Added `Changelog` to check latest version response.
 - Added `Critical` parameter.
+- Added `Create, List, Delete` App functionality. It now works like `Platforms, Channels, Archs`.
 
 ### Maintenance
 
 - Refactored `List`, `Create`, `Delete` mongodb functions.
-- `Platforms`, `Channels`, `Archs` was moved to `apps_meta` collection. 
+- `Platforms`, `Channels`, `Archs` have been moved to the `apps_meta` collection. 
 
 ## v1.3.3